+++ code/Meerkat.Caching/Caching/MemoryObjectCache.cs
@@ -13,6 +13,11 @@ namespace Meerkat.Caching
         private readonly MemoryCache cache;
         private readonly IRegionKeyStrategy keyStrategy;
 
+        /// <summary>
+        /// Creates a new instance of the <see cref="MemoryObjectCache"/> class.
+        /// </summary>
+        /// <param name="cache"></param>
+        /// <param name="keyStrategy"></param>
         public MemoryObjectCache(MemoryCache cache, IRegionKeyStrategy keyStrategy)
         {
             this.cache = cache;
@@ -35,16 +40,19 @@ namespace Meerkat.Caching
             }
         }
 
+        /// <copydoc cref="ICache.CacheMemoryLimit" />
         public long CacheMemoryLimit
         {
             get { return cache.CacheMemoryLimit; }
         }
 
+        /// <copydoc cref="ICache.Name" />
         public string Name
         {
             get { return cache.Name; }
         }
 
+        /// <copydoc cref="ICache.AddOrGetExisting" />
         public object AddOrGetExisting(string key, object value, DateTimeOffset absoluteExpiration, string regionName = null)
         {
             // Just do a get if we are null as MemoryCache doesn't store nulls.
@@ -57,12 +65,15 @@ namespace Meerkat.Caching
             return cache.AddOrGetExisting(regionKey, value, absoluteExpiration);
         }
 
+        /// <copydoc cref="ICache.Contains" />
         public bool Contains(string key, string regionName = null)
         {
             var regionKey = keyStrategy.Key(key, regionName);
 
             return cache.Contains(regionKey);
         }
+
+        /// <copydoc cref="ICache.Get(string, string)" />
 
         public object Get(string key, string regionName = null)
         {
@@ -71,11 +82,15 @@ namespace Meerkat.Caching
             return cache.Get(regionKey);
         }
 
+        /// <copydoc cref="ICache.Get" />
+
         public long GetCount(string regionName = null)
         {
             // NB This is not region aware
             return cache.GetCount();
         }
+
+        /// <copydoc cref="ICache.GetValues" />
 
         public IDictionary<string, object> GetValues(IEnumerable<string> keys, string regionName = null)
         {
@@ -84,12 +99,16 @@ namespace Meerkat.Caching
             return cache.GetValues(regionKeys);
         }
 
+        /// <copydoc cref="ICache.Remove" />
+
         public object Remove(string key, string regionName = null)
         {
             var regionKey = keyStrategy.Key(key, regionName);
 
             return cache.Remove(regionKey);
         }
+
+        /// <copydoc cref="ICache.Set" />
 
         public void Set(string key, object value, DateTimeOffset absoluteExpiration, string regionName = null)
         {