+++ code/Meerkat.Caching/Caching/MemoryObjectCacheFactory.cs
@@ -2,8 +2,15 @@
 
 namespace Meerkat.Caching
 {
+    /// <summary>
+    /// Factory for creating <see cref="MemoryObjectCache"/> with a region support
+    /// </summary>
     public static class MemoryObjectCacheFactory
     {
+        /// <summary>
+        /// Get a reference to the default <see cref="MemoryObjectCache"/> instance.
+        /// </summary>
+        /// <returns></returns>
         public static MemoryObjectCache Default()
         {
             var strategy = new RegionKeyStrategy();
@@ -12,6 +19,11 @@ namespace Meerkat.Caching
             return new MemoryObjectCache(cache, strategy);
         }
 
+        /// <summary>
+        /// Create a new instance of the <see cref="MemoryObjectCache"/>.
+        /// </summary>
+        /// <param name="name">Name to use</param>
+        /// <returns></returns>
         public static ICache Create(string name)
         {
             var strategy = new RegionKeyStrategy();